Chat_with_Your_PDF/requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
langchain==0.0.174
langchain-ollama==0.0.4
langchain-core==0.0.101
langchain-community==0.0.33
PyMuPDF==1.19.6